snippets/sdk/python/requirements.txt
azure-identity
azure-mgmt-authorization==3.0.0
azure-mgmt-keyvault==7.0.0
azure-keyvault-certificates==4.6.0
azure-mgmt-msi==6.1.0
azure-mgmt-nginx==2.0.0
azure-mgmt-network==21.0.1
azure-mgmt-resource==15.0.0
azure-mgmt-storage==16.0.0